backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rei Tsukada <rayzou1120@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.3"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
mypy = "^1.10.0"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.0.0"
sqlmodel = "^0.0.18"
pydantic-settings = "^2.2.1"
openai = "^1.30.1"
langchain = "^0.1.20"
pymupdf = "^1.24.4"
pypdf = "^4.2.0"
selenium = "^4.21.0"
bs4 = "^0.0.2"
mysqlclient = "^2.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
